remove the backend code for the part 2 (#7)
* add dist folder to ignore file

* remove package-lock.json from ignore file

This should always be commited with the code

* remove TinyMCE React package

The users following the article will need to add it themselves

* update backend package-lock.json

* remove article content

The users following the article will need to add those themselves

* Add the steps of the part 1 of the article

This make the TinyMCE part functional as a starting point for the part 2 of the series.

* remove the backend code for the part 2

The user will have to add this code themselves in the second part of the series.